requirements.txt
pip==23.0.1
setuptools==66.0.0
wheel==0.38.4
unidecode
numpy
pandas
sqlalchemy
python-dotenv
pymysql
